Enhance Next.js configuration and SEO features
- Updated next.config.ts to enable experimental features, optimize image formats, and add security headers for improved SEO.
- Added redirects for common misspellings to enhance user experience.
- Enhanced robots.txt to allow indexing for specific pages and added crawl-delay for better server performance.
- Improved metadata in layout.tsx for better SEO, including enhanced title, description, and keywords.
- Introduced sitemap.ts for structured data representation and improved search engine indexing.
- Created opengraph-image.tsx for dynamic Open Graph image generation.
- Refactored analytics page to separate logic into AnalyticsPageComponent for better organization and SEO optimization.
